pyproject.toml
[project]
name = "mlops-with-databricks"
version = "0.0.1"
description = "MLOps with Databricks"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = ["mlflow==2.17.0",
                "cffi==1.17.1",
                "cloudpickle==3.1.0",
                "lightgbm===4.5.0",
                "matplotlib==3.9.2",
                "numpy==1.26.4",
                "pandas==2.2.3",
                "psutil==6.0.0",
                "pyarrow==14.0.1",
                "scikit-learn==1.5.2",
                "scipy==1.14.1",
                "databricks-feature-engineering==0.6",
                "databricks-feature-lookup==1.2.0",
                "databricks-sdk==0.32.0"
                ]


[project.optional-dependencies]
dev = ["databricks-connect>=15.4.1, <16",
       "ipykernel>=6.29.5, <7",
       "pip>=24.2"]
